// apps/web/src/app/features/inventory/modules/drug-units/types/drug-units.types.ts
export interface DrugUnit {
  id: number;
  unit_code: string;
  unit_name: string;
  unit_name_en?: string | null;
  unit_type?: string | null;
  description?: string | null;
  is_active?: boolean | null;
  created_at?: string | null;
  updated_at?: string | null;
}